fix: make utils and constants treeshakeable
